fixes comments in PR and combines Password.php and DrupalPasswordChecker.php. Deletes Password.php
