+++ app/Services/Registrar.php
@@ -13,15 +13,24 @@ class Registrar
         $this->drupal_password_checker = $drupal_password_checker;
     }
 
-    public function login($user, $input, $login_type)
+    public function login($input)
     {
+        $login_type = 'username';
+        if ($input['email']) {
+            $email = strtolower($input['email']);
+            $user = User::where('email', '=', $email)->first();
+            $login_type = 'email';
+        } elseif ($input['mobile']) {
+            $user = User::where('mobile', '=', $input['mobile'])->first();
+            $login_type = 'mobile';
+        }
+
         if (($user instanceof User) && Hash::check($input['password'], $user->password)) {
             $token = $user->login();
             $token->user = $user->toArray();
 
             // Return the session token with the user.
             $user->session_token = $token->key;
-            $data = $user;
             return $user;
         } else if (($user instanceof User) && !($user->password)) {
 
@@ -38,7 +47,6 @@ class Registrar
 
                 // Return the session token with the user.
                 $user->session_token = $token->key;
-                $data = $user;
                 return $user;
             }
         } else {
